fix: cot moc thoi gian mili-giay tran INTEGER gay 500 khi gui tin nhan
now_vn_ts() tra ve epoch mili-giay (~1.78e12), vuot pham vi INTEGER 4-byte
cua Postgres (~2.1e9). Bang tin_nhan moi tao gan day dung dung Integer nen
loi ngay lap tuc; doi model sang BigInteger + them migration ALTER COLUMN
TYPE BIGINT cho tin_nhan.dau_moc_gui_ts, don_hang.dau_moc_tao,
lich_su_no.dau_moc (phong truong hop cac bang cu cung con INTEGER).

+++ backend/app/models/chat.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, ForeignKey, UniqueConstraint
+from sqlalchemy import BigInteger, Column, Integer, String, Text, ForeignKey, UniqueConstraint
 from sqlalchemy.orm import relationship
 from app.database import Base
 
@@ -38,7 +38,7 @@ class TinNhan(Base):
     noi_dung = Column(Text, nullable=True)
     loai_tin = Column(String, default="text")  # "text" | "anh" | "tep" | "he_thong"
     thoi_gian_gui = Column(String, default="")
-    dau_moc_gui_ts = Column(Integer, default=0)
+    dau_moc_gui_ts = Column(BigInteger, default=0)
 
     kenh_chat = relationship("KenhChat", back_populates="tin_nhans")
     tep_dinh_kems = relationship("TepDinhKem", back_populates="tin_nhan", cascade="all, delete-orphan")
